+++ src/main/java/team/incude/gsmc/v2/domain/auth/application/usecase/service/RefreshService.java
@@ -7,7 +7,7 @@ import team.incude.gsmc.v2.domain.auth.exception.RefreshTokenInvalidException;
 import team.incude.gsmc.v2.domain.auth.presentation.data.response.AuthTokenResponse;
 import team.incude.gsmc.v2.domain.member.application.port.MemberPersistencePort;
 import team.incude.gsmc.v2.domain.member.domain.Member;
-import team.incude.gsmc.v2.global.security.jwt.presentation.dto.TokenDto;
+import team.incude.gsmc.v2.global.security.jwt.dto.TokenDto;
 import team.incude.gsmc.v2.global.security.jwt.application.usecase.JwtIssueUseCase;
 import team.incude.gsmc.v2.global.security.jwt.application.usecase.JwtParserUseCase;
 import team.incude.gsmc.v2.global.security.jwt.application.usecase.JwtRefreshManagementUseCase;
